web/fix: title & favicon data response
